==> 1. Master File Generator.ipynb <==
import pandas as pd
import os

from os import walk

filenames = next(walk('/Users/thomasvanhoorn/Documents/THESIS NEW/DATA/1. MASTER FILE GENERATOR/archive'), (None, None, []))[2]

filenames.sort()

filenames = filenames[8:len(filenames)-1]

combined_years = pd.read_csv('/Users/thomasvanhoorn/Documents/THESIS NEW/DATA/1. MASTER FILE GENERATOR/archive/1996.QTR1.csv')
combined_years = combined_years[combined_years.form.isin(['10-K','10-Q'])]

with open('/Users/thomasvanhoorn/Documents/THESIS NEW/DATA/1. MASTER FILE GENERATOR/CONSTITUENT TICKER TEXT FILES/1996.txt') as f:
    lines = f.read().splitlines()

combined_years = combined_years[combined_years.form.isin([lines])]

combined_years['Y_Q'] = str('1996.QTR1')

for i in filenames:
    x = '/Users/thomasvanhoorn/Documents/THESIS NEW/DATA/1. MASTER FILE GENERATOR/archive/' + str(i)
    
    with open('/Users/thomasvanhoorn/Documents/THESIS NEW/DATA/1. MASTER FILE GENERATOR/CONSTITUENT TICKER TEXT FILES/' + str(i[0:4])+'.txt') as f:
        lines = f.read().splitlines()
    
    df = pd.read_csv(x)
    df = df[df.form.isin(['10-K','10-Q'])]
    df = df[df.symbol.isin(lines)]
    df['Y_Q'] = str(i[0:9])
    combined_years = pd.concat([combined_years,df])

combined_years = combined_years.reset_index(drop=True)

combined_years.to_csv(r'/Users/thomasvanhoorn/Documents//THESIS NEW/DATA/1. MASTER FILE GENERATOR/total.csv', index=False)
