==> 2. PREPARE MASTERFILE AND FACTOR FILE.ipynb <==
# Prepare file

import pandas as pd
import numpy as np
import os
import json
from bs4 import BeautifulSoup
import requests
from tqdm import tqdm
import math
import re
import time
import os, psutil

df = pd.read_csv('/Users/thomasvanhoorn/Documents/THESIS NEW/DATA/1. MASTER FILE GENERATOR/total.csv')

listje = list(df['cik'].unique())

with open("/Users/thomasvanhoorn/Documents/THESIS NEW/DATA/2.PREPARE MASTERFILE AND FACTOR FILE/INPUT FILES/allspciks.txt", 'w') as f:
    for item in listje:
        f.write("%s\n" % item)
        
f.close()

cik_cusip = pd.read_csv('/Users/thomasvanhoorn/Documents/THESIS NEW/DATA/2.PREPARE MASTERFILE AND FACTOR FILE/INPUT FILES/CIK CUSIP LINK.csv')
cik_cusip = cik_cusip.drop_duplicates(subset = ["cik"])
cik_cusip = cik_cusip.rename(columns={'cusip': 'cusip2'})

cik_cusip['cusip2'] = cik_cusip['cusip2'].str[:-1]

cik_cusip

listje = list(cik_cusip['cusip2'].unique())

with open("/Users/thomasvanhoorn/Documents/THESIS NEW/DATA/2.PREPARE MASTERFILE AND FACTOR FILE/INPUT FILES/allspcusips.txt", 'w') as f:
    for item in listje:
        f.write("%s\n" % item)
        
f.close()

cusip_permno = pd.read_csv('/Users/thomasvanhoorn/Documents/THESIS NEW/DATA/2.PREPARE MASTERFILE AND FACTOR FILE/INPUT FILES/CUSIPS.csv')


cusip_permno =  cusip_permno.drop_duplicates(subset = ["PERMNO"])

cusip_permno



cusip_permno = cusip_permno.rename(columns={'CUSIP': 'cusip2'})

cusip_permno = cusip_permno.reset_index(drop=True)

cusip_permno['cusip2'] = cusip_permno['cusip2'].str[:-1]

cusip_permno

linking_table = pd.merge(cik_cusip, cusip_permno, on=['cusip2'], how='inner')
linking_table

linking_table2 = linking_table[['gvkey','PERMNO','sic','cik','TICKER','cusip2']]
linking_table2 = linking_table2.rename(columns={'cusip2': 'cusip'})
linking_table2.to_csv('/Users/thomasvanhoorn/Documents/THESIS NEW/DATA/2.PREPARE MASTERFILE AND FACTOR FILE/linking table.csv')
linking_table2

df2 = pd.merge(df, linking_table2, on=['cik'], how='inner')

df2 = df2[['symbol', 'cik','form','access_number','filed_date', 'report_url','filing_url','Y_Q','sic','PERMNO']]




df2['filed_date'] = pd.to_datetime(df2['filed_date'])
df2['date'] = df2['filed_date'].apply(lambda x: x.date())
df2 = df2.rename(columns={'PERMNO': 'permno'})

# IMPORT FACTOR VARS AND ADD TO DF

import pyreadr

result = pyreadr.read_r('/Users/thomasvanhoorn/Documents/THESIS NEW/DATA/2.PREPARE MASTERFILE AND FACTOR FILE/INPUT FILES/Dataset_US.rds')

factor_vars = result[None]

factor_vars['permno'] = factor_vars['permno'].apply(lambda x: int(x))

listje = list(df2['permno'].unique())

listje = np.unique(listje)

factor_vars = factor_vars[factor_vars.permno.isin(listje)]

listje = list(factor_vars['permno'].unique())

listje = np.unique(listje)

df2 = df2[df2.permno.isin(listje)]

df2.to_csv("/Users/thomasvanhoorn/Documents/THESIS NEW/DATA/2.PREPARE MASTERFILE AND FACTOR FILE/INPUT FILES/final_index_file.csv")

# Merge the two files

factor_vars = factor_vars.reset_index(drop=True)

factor_vars['year'] = pd.DatetimeIndex(factor_vars['date']).year
df2['year'] = pd.DatetimeIndex(df2['date']).year


factor_vars['month'] = pd.DatetimeIndex(factor_vars['date']).month
df2['month'] = pd.DatetimeIndex(df2['date']).month


z = df2.groupby('permno')['year'].unique()
z =  pd.DataFrame(z)

z.index.name = 'newhead'
z.reset_index(inplace=True)

v = factor_vars.loc[factor_vars.permno==z['newhead'][0]][factor_vars.loc[factor_vars.permno==z['newhead'][0]].year.isin(z['year'][0])]

for i in range(1,len(z)):
    y = list(z['year'][i])
    
    v2 = factor_vars.loc[factor_vars.permno==z['newhead'][i]][factor_vars.loc[factor_vars.permno==z['newhead'][i]].year.isin(y)]
    
    v = pd.concat([v,v2], ignore_index = True)
    
factor_vars = v

factor_vars.to_csv("/Users/thomasvanhoorn/Documents/THESIS NEW/DATA/2.PREPARE MASTERFILE AND FACTOR FILE/INPUT FILES/factor_vars_final.csv")
