==> 6. SENTIMENT ANALYSIS 10KQ MDA.ipynb <==
!pip install pysentiment2
!pip install pickle5
!pip install transformers

from google.colab import drive
drive.mount('/content/drive')

import pandas as pd
import numpy as np
import nltk
nltk.download('punkt')
from nltk import tokenize
import re
import torch
from tqdm import tqdm
import random
import pysentiment2 as ps
from google.colab import files

#import 10KQ

unpickled_df = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/DATA/IDX_FILE_TEXT_DAT.csv')

unpickled_df = unpickled_df.reset_index(drop = True)

unpickled_df["negative"] = np.nan
unpickled_df["neutral"] = np.nan
unpickled_df["positive"] = np.nan

unpickled_df['Positive'] = np.nan
unpickled_df["Negative"] = np.nan
unpickled_df['Polarity'] = np.nan
unpickled_df['Subjectivity'] = np.nan

from transformers import AutoTokenizer, AutoModelForSequenceClassification
from transformers import pipeline

tokenizer = AutoTokenizer.from_pretrained("ProsusAI/finbert", model_max_len=512, truncation=True)

model = AutoModelForSequenceClassification.from_pretrained("ProsusAI/finbert")

device = torch.device("cuda") if torch.cuda.is_available() else torch.device("cpu")

model.to(device)

nlp = pipeline('sentiment-analysis',model=model, tokenizer=tokenizer, device = 0)




gpu_info = !nvidia-smi
gpu_info = '\n'.join(gpu_info)
if gpu_info.find('failed') >= 0:
  print('Not connected to a GPU')
else:
  print(gpu_info)

random.seed(3457)

dictio_finbert = {}
dictio_LM = {}

lm = ps.LM()

for i in tqdm(range(len(unpickled_df))):
  
  try:
      y = re.sub(r"[^a-zA-Z0-9 ]", " ", unpickled_df['text'][i])
      y = y.lower()
      y = re.sub(' {2,}', ' ', y)
      
      tokens = lm.tokenize(y)
      a = lm.get_score(tokens)   

      for r in a:
        unpickled_df.loc[i,r] = a[r]
      
      x = tokenize.sent_tokenize(unpickled_df['text'][i])
      
      x = [re.sub(r"[^a-zA-Z0-9 ]", " ", i) for i in x]

      x = [i.lower() for i in x]

      x = [re.sub(' {2,}', ' ', i) for i in x]
      
      x_new = []

      for o in range(len(x)):
          if len(x[o]) <= 1:
              continue
          elif len(x[o].split()) > 60:
              #x_new.append(x[o])
              continue     
          else:
              x_new.append(x[o])     

      z = []
      k = ''
      for p in x_new:
          if (len(tokenizer(k)[0]) + len(tokenizer(p)[0]))<150:
              p = ' ' + p
              k += p
          else:
              k = k[1:]
              if k != '':
                  if k[0] == ' ':
                      k = k[2:]
                      z.append(re.sub(' {2,}', ' ', k))
                      k = ''
                  else:
                      z.append(re.sub(' {2,}', ' ', k))   
                      k = ''
              else:
                  continue
          
      z = random.sample(z, int((len(z)/100)*100))
      
      label_list = []
      score_list = []
      
      for h in z:
          result = nlp(h)
          label = result[0]['label']
          score = result[0]['score']
          label_list.append(label)
          score_list.append(score)
          
      dictio_finbert[i] = [label_list,score_list]
              
      values, counts = np.unique(label_list, return_counts=True)

      ziped = zip(values, counts)

      ziped_list = list(ziped)

      #print(values)   
      #z = counts/sum(counts)
      print(list(ziped))


      if len(ziped_list) > 0:
        for c in ziped_list:

          unpickled_df.loc[i,c[0].lower()] = c[1]

      unpickled_df.loc[i, 'text'] = np.nan
      #print(unpickled_df['report_url'][i])
  except:
      continue


import pickle

pickle_out = open('/content/drive/MyDrive/Colab Notebooks/DATA/10kq_sentiment_new_prosusai_24911_37361.pkl', "wb")
pickle.dump(unpickled_df,pickle_out)
pickle_out.close()




import pickle

pickle_out = open('/content/drive/MyDrive/Colab Notebooks/DATA/output_6_11000_13068_10kq_sentiment.pkl', "wb")
pickle.dump(output_6,pickle_out)
pickle_out.close()


from google.colab import files

unpickled_df.to_pickle('output.pkl') 
files.download('output.pkl')
